@@ -9,7 +9,7 @@
 		}
 	],
 	"require": {
-		"php": ">=7.1.0",
+		"php": ">=7.4.0",
 		"ext-json": "*",
 		"ext-mbstring": "*",
 		"ext-iconv": "*",
@@ -21,7 +21,8 @@
 		"phpstan/phpstan": "^0.12.18",
 		"tracy/tracy": "^2.7",
 		"nette/security": "^3.0",
-		"phpstan/phpstan-nette": "^0.12.6"
+		"phpstan/phpstan-nette": "^0.12.6",
+		"symplify/easy-coding-standard": "^7.2"
 	},
 	"autoload": {
 		"classmap": [
